Add a raw_test_file filed to the component.
This PR adds a `raw_test_file` field to the component package object.
This allows for the inclusion of row text data. when raw text data is
included in this way, during the inline process, a config-map will be
used to store the raw data, with one-config map generated per file. Both
the metadata.name and the key in the data file is the base-file name of
the referenced file.