Switch to unstructured.Unstructured type.
This PR switches to Unstructured for representing unknown K8S objects,
which allows us to finish client generation.

This best corresponds to the contents and is compatible with the current
Proto schema and I think is a good semantic fit. We can't know the
schema of all the types, just that they have K8S-shape.