Add an AppVersion field to the ComponentSpec.
This PR adds AppVersion as a field to the component spec. Note that the
versioning is much less restrictive here than for the `version` field.
The app version will accept X.Y and X.Y.Z-blah.0 in addition to X.Y.Z

Note: This intentionally doesn't add requirements. I now see this more
as something that would live as a CRD in the component objects.