Default to not allowing unknown fields.
This PR updates ghodss and then disallows unknown fields for json/yaml
conversion, by default. Because the bundle definitions are typically
edited by hand, it's very easy to make typos that are hard to catch.
This should help the problem significantly.

Additionally, add some table tests for the muxer to test out this new case.

Fixes #121