Finish openapi validation method.
This method uses OpenAPI schemas to validate arbitrary JSON structures.
I moved this to an openapi directory from validate because this is also
where I will put defaulting logic. The openapi library takes the result
of a validate call and allows you to use it to other transformations
(e.g., defaulting)