Re-add the copied in conversion logic from apiextenisons-apiserver
Now that the cluster bundle is on K8S apimachinery >1.11, in theory we
can start using the conversion logic from 1.11 instead of this copied in
thing. Unfortunately, the internal library for apiextensions-apiserver
is going to need some work before the OpenAPI Schema validation can be
imported. So, to not block on that work, re-importing for the time
being.