filter.go: add a SelectObjects method
This adds SelectObjects. It works a little differently than the current
object selection because it returns the component with only the selected
objects. I think this is the cleanest method since it allows us to reuse
the ComponentPredicate that are already defined. I thought about only
making it return []*unstructured.Unstructured but I think the
information loss is not worth it. We can later write an
AllObjects([]*bundle.Component)[]*unstructured.Unstructured if we want
the same behavior as before.