filter.go: add generic filtering mechanism (#309)
This adds methods to select Components and Objects from a slice of Components. It allows custom predicates to determine the selection. There are methods to create a predicate that matches Component fields  as well as Object fields. It also provides And, Or, and Not methods to easily combine predicates. There are examples of Component selection and Object selection in the go doc https://pkg.go.dev/github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/filter?tab=doc.